backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# OCR and Computer Vision
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.9.0.80
numpy==1.24.3
Pillow==10.0.0

# Deep Learning and AI
# tensorflow-macos==2.15.0
tensorflow==2.15.0
protobuf>=3.20.3,<5.0.0dev
deepface==0.0.79
# dlib==19.24.2 --only-binary=all  # Removed - using OpenCV for face detection instead
scikit-image==0.21.0

# Text processing and matching
rapidfuzz==3.6.1
python-levenshtein==0.21.1

# Encryption and Security
cryptography==41.0.7
bcrypt==4.0.1

# File handling and cloud storage
python-dotenv==1.0.1
boto3==1.34.34
minio==7.2.3

# Task queue and caching
celery==5.3.6
redis==5.0.1
kombu==5.3.4

# Email and messaging
jinja2==3.1.2
aiohttp==3.9.1
aiosmtplib==3.0.1

# Blockchain integration
web3==6.15.1
eth-account==0.10.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
ruff==0.2.1
mypy==1.8.0
black==23.12.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0 